use runInBand instead of single test file
