+++ .github/workflows/updates-e2e.yml
@@ -95,7 +95,7 @@ jobs:
       - name: 🍏 Run iOS tests
         timeout-minutes: 30
         working-directory: packages/expo-updates
-        run: yarn test --config e2e/jest.config.ios.js
+        run: yarn test --config e2e/jest.config.ios.js --runInBand
       - name: 🤖 Start Android emulator and run tests
         timeout-minutes: 30
         uses: reactivecircus/android-emulator-runner@v2
@@ -104,5 +104,5 @@ jobs:
           avd-name: avd-${{ matrix.api-level }}
           arch: x86_64
           force-avd-creation: false
-          script: yarn test --config e2e/jest.config.android.js
+          script: yarn test --config e2e/jest.config.android.js --runInBand
           working-directory: packages/expo-updates
