--- packages/expo-updates/e2e/__tests__/Updates-e2e.test.ts
@@ -1,7 +1,0 @@
-import basic from './tests/basic';
-
-// run test suites from a single file in order to keep them entirely sequential;
-// --runInBand is not enough on its own
-describe('Updates e2e', () => {
-  basic();
-});
